+++ Vault/src/utopia/vault/nosql/access/single/model/distinct/SingleIdModelAccess.scala
@@ -1,7 +1,6 @@
 package utopia.vault.nosql.access.single.model.distinct
 
 import utopia.flow.datastructure.immutable.Value
-import utopia.vault.database.Connection
 import utopia.vault.nosql.factory.FromResultFactory
 
 /**
@@ -13,7 +12,4 @@ class SingleIdModelAccess[+A](val id: Value, val factory: FromResultFactory[A])
 	extends UniqueModelAccess[A]
 {
 	override def condition = table.primaryColumn.get <=> id
-	
-	// Generally it's unnecessary to request id based on the id itself
-	override def index(implicit connection: Connection) = id
 }